State Machine Addition
Major overhaul to the logic of the game. Instead of the root node being the game, it is now a state machine which runs on either a menu or playing state. The menu state machine will contain all current and future menus. Currently, the gameplay state machine is slightly underwhelming in its functionality since a lot of gameplay logic is found in the ball scene, which is almost certainly subject to change. This will hopefully allow the game to be much more scalable and make tasks such as adding game over end screens and new menus much easier.